feat: Refactor events module routing into presenter-based modes
