+++ plugins/event-list/templates/plugin.event-list.tpl
@@ -7,13 +7,13 @@
                 <li class="list-group-item">
                     <h5>
                         {$event.dateTimePeriod}<br />
-                        <a href="{$urlAdmidio}/modules/events/events.php?view_mode=html&view=detail&dat_uuid={$event.uuid}">{$event.headline}</a>
+                        <a href="{$urlAdmidio}/modules/events.php?mode=cards&dat_uuid={$event.uuid}">{$event.headline}</a>
                     </h5>
                     <div>{$event.description}</div>
                 </li>
             {/foreach}
             <li class="list-group-item">
-                <a href="{$urlAdmidio}/modules/events/events.php">{$l10n->get('PLG_EVENT_LIST_ALL_EVENTS')}</a>
+                <a href="{$urlAdmidio}/modules/events.php">{$l10n->get('PLG_EVENT_LIST_ALL_EVENTS')}</a>
             </li>
         </ul>
     {else}
